# fix(security): Deploy critical security fixes from PRs #200-203 with performance optimizations