--- a/src/vibe_check/server.py
+++ b/src/vibe_check/server.py
@@ -41,8 +41,18 @@ except ImportError:
         sys.exit(1)
 
 # Apply security patches to mcp_sampling (Issue #194)
-from .mentor.mcp_sampling_patch import auto_apply
-auto_apply()
+# Control via VIBE_CHECK_SECURITY_PATCHES environment variable
+# Default: enabled (patches active for security)
+# To disable: VIBE_CHECK_SECURITY_PATCHES=false
+SECURITY_PATCHES_ENABLED = os.getenv('VIBE_CHECK_SECURITY_PATCHES', 'true').lower() != 'false'
+
+if SECURITY_PATCHES_ENABLED:
+    from .mentor.mcp_sampling_patch import auto_apply
+    auto_apply()
+    # Log security status after logger is configured
+else:
+    # Security patches explicitly disabled
+    pass
 
 from .tools.analyze_text_nollm import analyze_text_demo
 from .tools.large_prompt_demo import demo_large_prompt_analysis
@@ -73,6 +83,14 @@ logging.basicConfig(
     ]
 )
 logger = logging.getLogger(__name__)
+
+# Log security patch status (Issue #194)
+if SECURITY_PATCHES_ENABLED:
+    logger.info("✅ Security patches ACTIVE (0.02% overhead) - 12 vulnerabilities patched")
+    logger.info("   Control via VIBE_CHECK_SECURITY_PATCHES environment variable")
+else:
+    logger.warning("⚠️ Security patches DISABLED - 12 vulnerabilities exposed!")
+    logger.warning("   Set VIBE_CHECK_SECURITY_PATCHES=true to enable protection")
 
 # Initialize FastMCP server
 mcp = FastMCP(