… benchmarks/performance_optimization_report.md
… benchmarks/security_performance_benchmark.py
… docs/SECURITY.md
… src/vibe_check/mentor/mcp_sampling_optimized.py
… tests/security/README_SECURITY_TESTS.md
… tests/security/test_security_regression.py
… tests/unit/test_educational_content.py