@kesslerio 2025-08-12
## ✅ PR Updated with Safety Mechanism

Based on the assessment feedback, I've added the recommended environment flag for safe deployment:

### Changes Made:
1. **Added `VIBE_CHECK_SECURITY_PATCHES` environment variable**
   - Default: `true` (patches enabled - safe with 0.02% overhead)
   - Emergency rollback: Set to `false` to disable

2. **Updated `server.py`** with conditional patch loading
   - Clear logging of security status on startup
   - Warning messages when patches are disabled

3. **Created `docs/SECURITY.md`** with comprehensive documentation
   - Configuration instructions
   - List of vulnerabilities fixed
   - Rollback procedures
   - Verification steps

### Why This is Now Safe to Merge:

✅ **Performance**: 0.02% overhead (0.008ms) is negligible
✅ **Safety**: Environment flag provides instant rollback
✅ **Default-On**: Maximum security out of the box
✅ **Simplicity**: No complex staging or feature flags needed
✅ **Testing**: All security and performance tests pass

The dramatic performance improvement (1,510x better than original) makes the complex phased deployment from Issue #204 unnecessary. This PR now provides both immediate security and a safety mechanism.

**Ready for merge and deployment.**