@kesslerio 2025-08-12
## 🎯 Simplification Complete - Ready to Merge

Just pushed the post-deployment cleanup as discussed:

### What I Did (Simple Version):
1. **Deleted the mess** - Moved 6 mcp_sampling variants to `deprecated/`
2. **Kept the fast one** - Using ultrafast version as main `mcp_sampling.py`
3. **Removed complexity** - No more runtime patching, no more environment flags
4. **Added one checkbox** - PR template now asks: "Is your code actually being used?"

### The Numbers:
- **Files**: 7 → 1 (85% reduction)
- **Performance**: Still 0.02% overhead
- **Security**: All 12 vulnerabilities still fixed
- **Complexity**: Dramatically reduced

### The Lesson Applied:
Instead of creating elaborate infrastructure for the cleanup, I just:
- Moved files (5 minutes)
- Updated one import (1 minute)
- Added one checkbox (30 seconds)

No frameworks. No automation. No quarterly audits. Just common sense.

**This PR is now even simpler and still fixes all the security issues. Ready to merge.**